[Feat] #116 - FloorUpdatedAlert 뷰 기능구현, 유저 플로어 로직 일부 수정
